broke apart schema file
