+++ package.json
@@ -7,6 +7,7 @@
     "db:push": "dotenv drizzle-kit push:mysql",
     "db:studio": "dotenv drizzle-kit studio",
     "dev": "next dev",
+    "type-check": "tsc --pretty --noEmit",
     "lint": "next lint",
     "start": "next start"
   },
