+++ src/server/api/actions/lessons.ts
@@ -1,7 +1,7 @@
 "use server";
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
 import { db } from "@/server/db";
-import { type CreateLesson, lessons } from "@/server/db/schema";
+import { type CreateLesson, lessons } from "@/server/db/schema/lessons";
 import { eq } from "drizzle-orm";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
