+++ src/server/api/actions/units.ts
@@ -1,7 +1,7 @@
 "use server";
 import { action } from "@/lib/actions";
 import { db } from "@/server/db";
-import { createUnitSchema, units } from "@/server/db/schema";
+import { createUnitSchema, units } from "@/server/db/schema/units";
 import { revalidatePath } from "next/cache";
 
 const createUnit = action(createUnitSchema, async (schema) => {
