+++ src/server/db/schema/index.ts
@@ -1,0 +1,8 @@
+export * from "./auth";
+export * from "./course";
+export * from "./lessons";
+export * from "./permissions";
+export * from "./subjects";
+export * from "./todo";
+export * from "./units";
+export * from "./table";
