+++ src/server/db/schema/table.ts
@@ -1,0 +1,10 @@
+/**
+ * This is an example of how to use the multi-project schema feature of Drizzle ORM. Use the same
+ * database instance for multiple projects.
+ *
+ * @see https://orm.drizzle.team/docs/goodies#multi-project-schema
+ */
+
+import { mysqlTableCreator } from "drizzle-orm/mysql-core";
+
+export const mysqlTable = mysqlTableCreator((table) => `creek-ocw_${table}`);
