+++ src/server/permissions/index.ts
@@ -1,5 +1,9 @@
-export const permissions = [
-  "basics",
+export const permissions = ["basics", "activites"] as const;
+
+export type Permission = (typeof permissions)[number];
+export type Permissions = Permission[];
+
+export const course_permissions = [
   "read:users",
   "add:lessons",
   "add:tags",
@@ -7,8 +11,11 @@ export const permissions = [
   "edit:lessons",
   "edit:tags",
   "edit:courses",
-  "activites",
 ] as const;
 
-export type Permission = (typeof permissions)[number];
-export type Permissions = Permission[];
+export type CoursePermission = (typeof course_permissions)[number];
+export type CoursePermissions = CoursePermission[];
+
+export const subject_permissions = ["read:users", "add:courses"];
+export type SubjectPermission = (typeof subject_permissions)[number];
+export type SubjectPermissions = SubjectPermission[];
